fix: coerce snapshot_to_series output to float
astype(float) after to_numeric guarantees float type dtype reagardless of
integer-value input strings

+++ src/phase_2/fred_data.py
@@ -41,7 +41,7 @@ def collapse_to_snapshot(releases):
 
 def snapshot_to_series(snap):
     series = pd.Series(
-        pd.to_numeric(snap['value'], errors='coerce').values,
+        pd.to_numeric(snap['value'], errors='coerce').astype(float).values,
         index=pd.to_datetime(snap['date']).values
     ).sort_index()
     return series
